Refactoring of device-plugin project structure
* Moved device-plugin and device monitor to separate packages
* Moved main.go to dedicated cmd folder
* Updated docker build container
