+++ cmd/k8s-device-plugin/k8s-device-plugin.go
@@ -5,21 +5,24 @@ package main
 
 import (
 	"flag"
-	"os"
 	"github.com/golang/glog"
+	"k8s-ne-device-plugin/pkg/nitro_enclaves_device_monitor"
+	"k8s-ne-device-plugin/pkg/nitro_enclaves_device_plugin"
+	"os"
 )
 
 func main() {
 	flag.Parse()
 	glog.V(0).Info("Loading K8s Nitro Enclaves device plugin...")
 
-	devicePlugin := NewNitroEnclavesDevicePlugin()
-	monitor := NewNitroEnclavesMonitor(devicePlugin)
 
+	// create nitro enclave device, pass it to monitor and start in background
+	devicePlugin := nitro_enclaves_device_plugin.NewNitroEnclavesDevicePlugin()
+	monitor := nitro_enclaves_device_monitor.NewNitroEnclavesMonitor(devicePlugin)
 	if monitor == nil {
 		glog.Error("Error while initializing NE plugin monitor!")
 		os.Exit(1)
 	}
-	
+
 	monitor.Run()
 }
